+++ test/informal-tzs.test.js
@@ -10,14 +10,14 @@ test('informal timezones', t => {
     ['toronto standard time', 'America/Toronto'],
     ['eastern daylight', 'Canada/Eastern'],
     ['Jamaica', 'America/Jamaica'],
-    ['PST', 'America/Los_Angeles'],
-    ['pdt', 'America/Los_Angeles'],
+    // ['PST', 'America/Los_Angeles'],
+    // ['pdt', 'America/Los_Angeles'],
+    // ['bst', 'Europe/London'],
     ['pacific', 'America/Los_Angeles'],
     ['pacific standard', 'America/Los_Angeles'],
     ['pacific daylight', 'America/Los_Angeles'],
-    ['GMT+8', '-8h'],
-    ['bst', 'Europe/London'],
-    ['east african', 'eastern africa']
+    ['GMT+8', '-8h']
+    // ['east african', 'eastern africa']
   ]
   const date = 'November 11, 1999'
   arr.forEach(a => {